+++ CHANGELOG.md
@@ -11,11 +11,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 - Add project option.
 - Remove project option.
+- List projects option.
 
 ### Updated
 
 - Help option with new add option information.
 - Help option with new remove option information.
+- Help option with new list option information.
 
 ## [0.3.0] - 2021-01-09